RNN model/requirements.txt
numpy>=1.13.1
scikit-learn>=0.18.1
scipy>=0.19.0
Pytorch>=0.4.1
torchtext>=0.1.1
